+++ resources/src/components/Icons.vue
@@ -1,0 +1,25 @@
+<template>
+  <ul class="icons">
+    <li>
+      <a href="https://twitter.com/cachilders" target="_blank">
+        <img src="../assets/img/twitter.svg">
+      </a>
+    </li>
+    <li>
+      <a href="https://github.com/cachilders/tender" target="_blank"><img src="../assets/img/github.svg"></a>
+    </li>
+  </ul>
+</template>
+
+<script>
+export default {
+  data() {
+    return {};
+  },
+};
+</script>
+
+<!-- Add "scoped" attribute to limit CSS to this component only -->
+<style scoped>
+
+</style>