+++ resources/src/views/About.vue
@@ -4,18 +4,16 @@
       <img src="../assets/img/logo.svg">
       <h1>TENDER</h1>
       <h2>Well, there's not much to say, here.</h2>
-    </div>
-    <div class="__footer">
-      <ul>
-        <li>
-          <a href="https://twitter.com/cachilders" target="_blank">
-            <img src="../assets/img/twitter.svg">
-          </a>
-        </li>
-        <li>
-          <a href="https://github.com/cachilders/tender" target="_blank"><img src="../assets/img/github.svg"></a>
-        </li>
-      </ul>
+      <icons></icons>
     </div>
   </div>
 </template>
+
+<script>
+import Icons from '../components/Icons.vue';
+export default {
+  components: {
+    Icons
+  },
+};
+</script>